GrowflowAi/requirements.txt
cryptography>=42.0.0
Flask>=3.0.0
Flask-Bcrypt>=1.0.1
Flask-Cors>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-Limiter>=3.8.0
Flask-SQLAlchemy>=3.1.1
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
qrcode>=7.4.2
requests>=2.31.0
